Robot base frame tracking is live
